Split SearchBox view into separate file
This allows us to write a basic sanity check for the SearchBox view.
